requirements.txt
python-dotenv
python-telegram-bot
twitchio
websockets
httpx